CryptoPkg/TlsLib: Remove NULL cipher
The term "NULL" refers to NULL-MD5, NULL-SHA and NULL-SHA256 when
used to set the SSL cipher list.  As both MD5 and SHA variants are
explicitly listed in our code, I surmise enabling all three by
setting the cipher list to just NULL was not the intended behavior.
This patch will remove NULL as an option for the cipher list and
allow NULL-SHA256 instead.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Thomas Palmer <thomas.palmer@hpe.com>
Reviewed-by: Samer El-Haj-Mahmoud <elhaj@hpe.com>
Reviewed-By: Wu Jiaxin <jiaxin.wu@intel.com>