StandaloneSmmPkg: Workaround incorrect Hoblist initialisation
The invocation of HobLibConstructor() is notional. It searches for a
pre-installed HobList in a configuration table in the SMST. However,
the installation does not take place until much later in
SmmMainStandalone(). To make the constructor initialisation
worthwhile, the reference to the Hoblist is passed in the ImageHandle
parameter of the constructors. The HobLibConstructor() uses it to
initialise its internal pointer.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Achin Gupta <achin.gupta@arm.com>