StandaloneSmmPkg: Get NS communication buffer details from Hoblist
On ARMv8-A platforms, synchronous communication between the MM
environment in the secure world and the UEFI runtime environment or
the OS in the normal world will take place through the MM_COMMUNICATE
SMC. A parameter of this SMC is a communication buffer. It is possible
for the caller to specify a different buffer for every call that would
have to subsequently mapped into the secure world.

However, it is more optimal if the secure world pre-allocated a buffer
and provided it extents to both the normal world and the MM
environment. This approach replaces the overhead of mapping a buffer
into the secure world translation tables potentially for every
MM_COMMUNICATE SMC with a simple bounds check of the buffer.

The software entity responsible for initialising the MM Foundation
provides the details of this buffer in one of the
EFI_SMRAM_DESCRIPTORs in the EFI_SMRAM_HOB_DESCRIPTOR_BLOCK. The
EFI_SMRAM_STATE for this SMRAM range is marked with a non-standard
attribute. This patch adds support for retrieving the extents of this
buffer from the Hoblist and making a private copy of this information.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Achin Gupta <achin.gupta@arm.com>