StandaloneSmmPkg: Add protocol to register handlers for GUIDed events
In the traditional mode, handlers for MM events are identified both
with and without GUIDs. The latter approach is used for asynchronous
MMIs where it is not possible for the caller to provide a GUID that
can be used to identify the event. As a result each registered handler
without an associated GUID is invoked until one or more of them handle
the event. The presence of a communication buffer in gSmmCorePrivate
array is used to distinguish between the two approaches.

On ARM platforms, there is enough information to identify a event
(SMC, Interrupt ID, Aborts) at the time of its occurrence. This makes
it possible to call the event specific handler directly instead of
invoking multiple handlers. ARM Trusted Firmware is responsible for
receiving an event and associating a unique event ID with it. MM
drivers use GUIDs to identify handlers for MM events. The CPU driver
is responsible for mapping event IDs to GUIDs and vice-versa. Hence,
it is reasonable to assume that on ARM platforms, handlers for both
synchronous and asynchronous events can be registered and identified
using GUIDs.

This patch introduces a new protocol that interfaces with ARM Trusted
Firmware to enable registration, unregistration and handling of GUIDed
events. MMST services are used to register and unregister handlers as
before. Additionally, the driver also looks up the event ID
corresponding to a GUID and registers/unregisters a handler with ARM
Trusted Firmware to enable it to delegate the event. Also, while
handling an event and interface is made available to the MM entry
point to obtain the GUID (in a MM_COMMUNICATE_HEADER) for that
event. This GUID can be used as a parameter to SmiManage() to handle
the event.

A dummy assembler entry point has also been added that:

1. Calls the C entrypoint of the CPU driver
2. Issues an SVC to return execution to ARM Trusted Firmware upon
   exit from 1.

The C entrypoint currently returns successfully and will be populated
in later patches.

The association between an event ID and a GUID could vary across
platforms. The implementation currently assumes that this is not the
case.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Achin Gupta <achin.gupta@arm.com>