StandaloneSmmPkg: Demonstrate use of GUIDed event handler
This patch demonstrates use of the GUIDed event management
protocol. It registers the SmmUefiInfoHandler() with the corresponding
GUID using the GuidedEventHandlerRegister() interface. The GUID is
matched to the SMC event id. The normal world is expected to invoke
the MM_COMMUNICATE SMC with the same GUID populated in the
communication buffer. The CPU driver copies the communication buffer
into the PerCpuGuidedEventContext entry for that
CPU. SmmUefiInfoHandler() retrieves this buffer to obtain a pointer to
the UEFI system table.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Achin Gupta <achin.gupta@arm.com>