StandaloneSmmPkg: Add ArmSvcLib
This patch adds a library that enables invocation of SVCs from EL0. It
will be used by the Standalone MM environment to request services from
a software running in a privileged EL e.g. ARM Trusted Firmware. The
library is a direct copy of the Arm SMC Library.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Achin Gupta <achin.gupta@arm.com>