StandaloneSmmPkg: Add EFI_MM_COMMUNICATION_PROTOCOL DXE driver
This patch implements the EFI_MM_COMMUNICATION_PROTOCOL DXE runtime
driver for AARCH64 platforms. It uses SMCs allocated from the standard
SMC range in the SMC Calling Convention specification to communicate
with the standalone MM environment in the secure world. Two SMCs are
used currently to:

1. Obtain the extents of the communication buffer to use for passing
   parameters
2. Request services from the Standalone MM environment in the secure world

If 1. does not work, then the driver relies on platform provided values to
determine the extents of the communication buffer.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Achin Gupta <achin.gupta@arm.com>