ArmPlatformPkg: Increment VE RTSM mem map descriptor count
Memory allocated for the MM communication buffer between the normal and
secure worlds is mapped in as a separate descriptor during MMU
initialisation. This patch increases the number of descriptors by 1 to
accommodate the MM buffer descriptor to avoid an assertion failure or a
possible memory corruption.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Achin Gupta <achin.gupta@arm.com>