ArmStdSmc: Rationalise SMC and SVC function identifiers for MM
The existing function identifiers for accessing Management Mode services are not
compliant with the MM Interface Specification. This patch fixes this issue.

Also, function identifiers used by the MM environment to interface with ARM
Trusted Firmware have been moved to a different range (0x60-0x7f). These changes
are not final but this patch updates the code to enable progress on the
prototyping front.

Lastly, function identifiers that are not being used or are likely to be
deprecated have been removed to minimise code pollution.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Achin Gupta <achin.gupta@arm.com>