PeCoffExtraActionLib: Add library to change PE-COFF section permissions
The Standalone MM environment runs in S-EL0 in AArch64 on ARM Standard
Platforms. The heap is assigned RW+XN attributes in the S-EL1/EL0
translation regime. MM drivers are relocated to the heap during
dispatch. Prior to execution, the sections of each driver have to be
assigned correct memory attributes to ensure correct operation. This
patch implements an PeCoffExtraActionLib that calls functions
implemented by the ArmMmuUnprivilegedLib to change the permissions of
these sections after the driver has been relocated.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Achin Gupta <achin.gupta@arm.com>