HobLib: Add a common HOB Library for producers and consumers
All HOB library implementations in EDK2 assume that they will be invoked
in either the producer or consumer phase but not both. In the former
case only functions for creating HOBs are implemented. In the latter
case only functions for parsing a HOB list are implemented.

The Standalone MM environment is initialised during the SEC phase on ARM
Standard Platforms. The MM Core driver implements an entry point module
which is architecture specific and runs prior to the generic core driver
code. The former creates a Hob list that the latter consumes. This
happens in the same phase.

This patch implements a Hob library that can be used by the entry point
module to produce a Hob list and by the core driver code to consume it.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Achin Gupta <achin.gupta@arm.com>