StandaloneSmmPkg: Export ARM TF CPU driver entry point
The AArch64 entry point driver needs to know the CPU driver entry point
in order to invoke it to delegate an event received from ARM Trusted. It
creates a HOB with a pointer that the CPU driver is expected to populate
with its entry point. This patch implements the support in the CPU
driver to find this HOB and populate the entry point address.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Achin Gupta <achin.gupta@arm.com>