StandaloneSmmPkg: Modify logic for handling delegated events in MM
This patch deprecates untested logic for registering and handling
asynchronouse events and refines the mechanism for handling synchronous
events.

For the latter, the use of the event registration SVC interface with ARM
TF is deprecated. The MM environment has a single entry point
(implemented by the Entry point driver) which in turn invokes the CPU
Driver entry point. The handler for a synchronous MM request (issued
through MM_COMMUNICATE) is identified using a GUID. Hence registering a
handler with ARM TF for synchronous events is redundant.

The untested logic to register and handle asynchronous events has been
removed and the relevant source files are not included in the
build. This will be revisted in the future.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Achin Gupta <achin.gupta@arm.com>