MpyTestFrameworkPkg: use minified jquery source
We should be using the compressed "minified" jquery source file
rather than the uncompressed version, as it will reduce the page
load times.

This updates the relevant FreeMaker templates, the readme, and
the setup script. It also corrects a small mistake in the error
text of the setup script.

Contributed-under: TianoCore Contribution Agreement 1.1.
Signed-off-by: Stephano Cetola <stephano.cetola@linux.intel.com>
Reviewed-by: Long Qin <qin.long@intel.com>