Announce to create "UEFI_Redfish" branch in edk2-staging.
v2: Resend the patch as diff adding instead of modifying.

UEFI_Redfish branch is to develop the UEFI Redfish feature. The code base
of development is based on the release of edk2-stable201811 tag. Please
refer to the patch of Readme.md to get the detailed feature introduction.

Note: The branch will be created by the end of Jan 28th if no objection.

Cc: Leif Lindholm <leif.lindholm@linaro.org>
Cc: Rothman Michael A <michael.a.rothman@intel.com>
Cc: Kinney Michael D <michael.d.kinney@intel.com>
Cc: Li Ruth <ruth.li@intel.com>
Cc: Ye Ting <ting.ye@intel.com>
Cc: Fu Siyuan <siyuan.fu@intel.com>
Cc: Wang Fan <fan.wang@intel.com>
Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Wu Jiaxin <jiaxin.wu@intel.com>