OvmfPkg/PlatformBootManagerLib: Process UEFI Capsules
Add logic to process UEFI Capsules before and after the
system lock point.  This change moves the lock point
after consoles are initialized if boot mode is
BOOT_OB_FLASH_UPDATE.

NOTE: Due to change in lock location, this change may
not be safe.  It assumes that the console devices for
QEMU are trusted consoles.  This supports testing of
capsule update for a platform that does have a trusted
console.  If lock point is performed before consoles,
then a different progress indicator is required that
does not use consoles(e.g. audio/beep codes).

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Michael D Kinney <michael.d.kinney@intel.com>