IntelSiliconPkg PeiFirmwareInterfaceTableLib: Add initial library (CVE-2019-11098)
REF:https://bugzilla.tianocore.org/show_bug.cgi?id=1614

Add a Firmware Interface Table (FIT) library that is used to
manage the FIT. The current interface allows the FIT base to
be acquired in a manner without boot media assumptions
(e.g. SPI flash MMIO) and a FIT HOB to be produced that can
make the FIT contents available without access to the SPI flash.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Michael Kubacki <michael.a.kubacki@intel.com>