MdeModulePkg/EsrtFmpDxe: Detect duplicate GUID/HardwareInstance
https://bugzilla.tianocore.org/show_bug.cgi?id=1525

Fix the following issues identified in #1525:
    EsrtFmpDxe - The check for more than one instance of the same
    GUID/HardwareInstance is not being done at the right point.
    It is being done as EFI_FIRMWARE_IMAGE_DESCRIPTORs are being merged
    into an ESRT Table Entry.  This means a case will be missed if
    there are 3 EFI_FIRMWARE_IMAGE_DESCRIPTORs with the first and 3rd
    one being having the same GUID/HardwareInstance.  Instead, this check
    should be performed across all EFI_FIRMWARE_IMAGE_DESCRIPTORs in the
    entire handle database.

Build a table of GUID/HardwareInstance pairs from all the
EFI_FIRMWARE_IMAGE_DESCRIPTORs from all FMP instances.  If a duplicate
is found, then generate a DEBUG_ERROR message, generate an ASSERT(),
and ignore the duplicate EFI_FIRMWARE_IMAGE_DESCRIPTOR.

Add an internal worker function called FmpGetFirmwareImageDescriptor()
that retrieves the list of EFI_FIRMWARE_IMAGE_DESCRIPTORs from a single
FMP instance and returns the descriptors in an allocated buffer.  This
function is used to get the descriptors used to build the table of
unique GUID/HardwareInstance pairs.  It is then used again to generate
the ESRT Table from all the EFI_FIRMWARE_IMAGE_DESCRIPTORs from all the
FMP instances.  2 passes are performed so the total number of
descriptors is known.  This allows the correct sized buffers to always
be allocated and the extra logic to grow tables is removed.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Eric Jin <eric.jin@intel.com>
Reviewed-by: Michael D Kinney <michael.d.kinney@intel.com>